% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_data}
\alias{create_data}
\title{Create data}
\usage{
create_data(filt_df)
}
\arguments{
\item{filt_df}{list with the data and flist objetcs}
}
\value{
data data.frame after filtering
}
\description{
Extract filtered data from filt_df object
}
\examples{
df <- data.frame(a = rnorm(100),
                          b = rnorm(100),
                          c = sample(letters[1:10], size = 100, replace = TRUE))

lf <- list("a < 1.96",
           "b < 1.96",
           "c \%in\% letters[1:5] | abs(b) > 0.2")

filt_df <- filter_list(data = df, lfilters = lf)
create_data(filt_df)

}
