% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_table0}
\alias{create_table0}
\title{Create table zero}
\usage{
create_table0(filt_df)
}
\arguments{
\item{filt_df}{list with the data and flist objetcs}
}
\value{
table0 data.frame with the number of study observations excluded with respect to each exclusion criteria
}
\description{
Extract from filt_df object the number of study observations excluded with respect to each exclusion criteria
}
\examples{
df <- data.frame(a = rnorm(100),
                          b = rnorm(100),
                          c = sample(letters[1:10], size = 100, replace = TRUE))

lf <- list("a < 1.96",
           "b < 1.96",
           "c \%in\% letters[1:5] | abs(b) > 0.2")

filt_df <- filter_list(data = df, lfilters = lf)
create_table0(filt_df)

}
