% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_list.R
\name{filter_list}
\alias{filter_list}
\title{Filter list}
\usage{
filter_list(data, lfilters)
}
\arguments{
\item{data}{data.frame to apply filters}

\item{list_filters}{list of filters to apply}
}
\value{
flist list with rows excluded by each filter

data data.frame after filtering
}
\description{
Filter a dataset given filters in a list
}
\examples{
df <- data.frame(a = rnorm(100),
                          b = rnorm(100),
                          c = sample(letters[1:10], size = 100, replace = TRUE))

lf <- list("a < 1.96",
           "b < 1.96",
           "c \%in\% letters[1:5] | abs(b) > 0.2")

filt_df <- filter_list(data = df, lfilters = lf)

}
